[refactor]: remove unused pages and components
- 기존 Desktop, Mobile 나뉜 페이지 하나로 통합 및 tailwind option 사용해서 간단하게 구현

- Deleted mobile and desktop components for Home, Grade, Graduation, Timetable, and Scholarship pages.
- Removed index files that exported these components.
- Updated router to directly import and use the new structure for pages.
- Cleaned up TypeScript configuration files to improve build performance.
