--- src/pages/grade/index.ts
@@ -1,2 +1,0 @@
-export { default as GradeMobile } from "./mobile";
-export { default as GradeDesktop } from "./desktop";
