--- src/pages/graduation/index.ts
@@ -1,2 +1,0 @@
-export { default as GraduationMobile } from "./mobile";
-export { default as GraduationDesktop } from "./desktop";
