--- src/pages/home/desktop.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const DesktopHomePage: React.FC = () => {
-  return (
-    <div className="min-h-screen bg-gray-50 flex items-center justify-center">
-      <div className="text-center">
-        <h1 className="text-4xl font-bold text-gray-900 mb-4">
-          건국대학교 학사정보시스템
-        </h1>
-        <p className="text-lg text-gray-600">
-          데스크톱 버전은 개발 중입니다.
-        </p>
-      </div>
-    </div>
-  );
-};
-
-export default DesktopHomePage;
