--- src/pages/home/index.ts
@@ -1,2 +1,0 @@
-export { default as HomeMobile } from "./mobile";
-export { default as HomeDesktop } from "./desktop";
