--- src/pages/index.ts
@@ -1,14 +1,0 @@
-// Home Pages
-export { HomeMobile, HomeDesktop } from "./home";
-
-// Grade Pages
-export { GradeMobile, GradeDesktop } from "./grade";
-
-// Graduation Pages
-export { GraduationMobile, GraduationDesktop } from "./graduation";
-
-// Timetable Pages
-export { TimetableMobile, TimetableDesktop } from "./timetable";
-
-// Scholarship Pages
-export { ScholarshipMobile, ScholarshipDesktop } from "./scholarship";
