--- src/pages/scholarship/index.ts
@@ -1,2 +1,0 @@
-export { default as ScholarshipMobile } from "./mobile";
-export { default as ScholarshipDesktop } from "./desktop";
