--- src/pages/timetable/index.ts
@@ -1,2 +1,0 @@
-export { default as TimetableMobile } from "./mobile";
-export { default as TimetableDesktop } from "./desktop";
